custom content generation

+++ sitebuilder/sitebuilder/management/commands/build.py
@@ -8,7 +8,7 @@ import shutil
 
 from django.conf import settings
 from django.core.management import call_command
-from django.core.management.base import BaseCommand
+from django.core.management.base import BaseCommand, CommandError
 from django.core.urlresolvers import reverse
 from django.test.client import Client
 
@@ -24,16 +24,28 @@ class Command(BaseCommand):
 
   def handle(self, *args, **options):
     """Request pages, generate content, output content"""
-    #if directory exists, wipe clean
-    if os.path.exists(settings.SITE_OUTPUT_DIRECTORY):
-      shutil.rmtree(settings.SITE_OUTPUT_DIRECTORY)
-    os.mkdir(settings.SITE_OUTPUT_DIRECTORY)
-    os.makedirs(settings.STATIC_ROOT)
+    #check is there is one or more args passed into command
+    if args:
+      pages = args
+      available = list(get_pages())
+      invalid = []
+      for page in pages:
+        if page not in available:
+          invalid.append(page)
+      if invalid:
+        msg = "Invalid pages: {}".format(', '.join(invalid))
+        raise CommandError(msg)
+    else:
+      #if directory exists, wipe clean
+      if os.path.exists(settings.SITE_OUTPUT_DIRECTORY):
+        shutil.rmtree(settings.SITE_OUTPUT_DIRECTORY)
+      os.mkdir(settings.SITE_OUTPUT_DIRECTORY)
+      os.makedirs(settings.STATIC_ROOT)
     #copy all static resources into STATIC_ROOT dir
     call_command('collectstatic', interactive=False,
         clear=True, verbosity=0)
     client = Client()
-    for page in get_pages():
+    for page in pages:
       #grabs all .html resources
       url = reverse('page', kwargs={'slug': page})
       response = client.get(url)
@@ -41,7 +53,8 @@ class Command(BaseCommand):
         output_dir = settings.SITE_OUTPUT_DIRECTORY
       else:
         output_dir = os.path.join(settings.SITE_OUTPUT_DIRECTORY, page)
-        os.makedirs(output_dir)
+        if not os.path.exists(output_dir):
+          os.makedirs(output_dir)
       #crawl thru .html pages and write content into
       #SITE_OUTPUT_DIRECTORY
       with open(os.path.join(output_dir, 'index.html'), 'wb') as f:
@@ -53,3 +66,6 @@ class Command(BaseCommand):
 
 #this will easily deploy hosting files to any hosting platform
 #for showing rapid prototype progress.
+
+#updated to build out single pages by kwarg
+#ex: python prototype.py build index
